Moving body parser middleware to server/bindBodyParser
