+++ lib/core/mount.js
@@ -33,8 +33,6 @@ var express = require('express');
 var favicon = require('serve-favicon');
 var morgan = require('morgan');
 var methodOverride = require('method-override');
-var multer = require('multer');
-var bodyParser = require('body-parser');
 var compression = require('compression');
 
 function mount(events) {
@@ -125,28 +123,14 @@ function mount(events) {
 		keystone.callHook('pre:bodyparser', req, res, next);
 	});
 
-	// Set up body options and cookie parser
-	var bodyParserParams = {};
-
-	if (this.get('file limit')) {
-		debug('adding file limit');
-		bodyParserParams.limit = this.get('file limit');
-	}
-
-	app.use(bodyParser.json(bodyParserParams));
-	bodyParserParams.extended = true;
-	app.use(bodyParser.urlencoded(bodyParserParams));
-	app.use(multer({
-		includeEmptyFields: true
-	}));
+	require('../../server/bindBodyParser')(this, app);
+	app.use(methodOverride());
 
 	// Add 'X-Frame-Options' to response header for ClickJacking protection
 	if (this.get('frame guard')) {
 		debug('enabling frame guard');
 		app.use(require('../security/frameGuard')(this));
 	}
-	
-	app.use(methodOverride());
 
 	// Pre route config
 	if ('function' === typeof this.get('pre:routes')) {
